Thinking

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -9,7 +9,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int count;
+	long unsigned int count;
 
 	if (!action || !array)
 		return;
